Add `resolve_vars_with_context` to find variable names in an image
Adds a new `ImageRef::resolve_vars_with_context()` to find all
variable names that were used to build an image while resolving the
end result.

Additionally adds a helper function, `Stage::arg_index()` that can
be used to determine if a stage contains an `ARG` instruction with
a particular name.

These two functions can be used together to determine if a particular
`ARG` is used only to parameterize an image in a `FROM`, or if the
content is actually used in some build stage.